Reduce duplicate code + speed up getitem
